requirements.txt
streamlit>=1.45.0
anthropic>=0.50.0
python-pptx>=1.0.2
pdfplumber>=0.11.4
plotly>=5.24.1
pandas>=2.2.3
